backend/app/services/user_service/pyproject.toml
[project]
name = "user_service"
version = "0.1.0"
description = "User microservice"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.30.0",
  "shared",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "pydantic[email]>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["user_service"]